+++ packages/zarm/src/__tests__/index.test.ts
@@ -465,9 +465,6 @@ describe('index', () => {
             "$$typeof": Symbol(react.forward_ref),
             "render": [Function],
           },
-          "defaultProps": Object {
-            "visible": true,
-          },
           "render": [Function],
         },
         "Tabs": Object {